enable multitenant, update errors
