+++ crates/ndc-graphql/src/main.rs
@@ -1,9 +1,9 @@
-use ndc_graphql::connector::GraphQLConnector;
+use ndc_graphql::connector::setup::GraphQLConnectorSetup;
 use ndc_sdk::default_main::default_main;
 
 use std::error::Error;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
-    default_main::<GraphQLConnector>().await
+    default_main::<GraphQLConnectorSetup>().await
 }
